test(maestro): refresh slot navigation gates
