+++ crates/vize_maestro/src/ide/definition.rs
@@ -1,4 +1,5 @@
 //! Go-to-definition for Vue SFC template bindings, components, imports, and Corsa.
+mod art;
 pub mod bindings;
 #[cfg(all(test, feature = "native"))]
 mod corsa_tests;
